+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# Changelog
+# CHANGELOG
 
+* Create the redirect method in controller
 * Save the slug
 * Handle errors on new view
 * Update README.me